refactor(schedule): extract responsive layouts into separate components
