+++ pages/conference/schedule.vue
@@ -13,55 +13,9 @@
             </schedule-day-tab>
         </div>
         <div class="font-serif">
-            <schedule-rooms class="lg:grid" :date-index="selectedDayIndex">
-                <span></span>
-                <schedule-room
-                    v-for="(room, i) in rooms"
-                    :key="$makeKey(i, 'schedule_room')"
-                    :value="room"
-                ></schedule-room>
-            </schedule-rooms>
-            <schedule-table :date-index="selectedDayIndex" class="lg:grid">
-                <schedule-tick
-                    v-for="(tick, i) in table.ticks"
-                    :key="$makeKey(i, 'schedule_tick')"
-                    :style="tick.style"
-                >
-                    {{ tick.label }}
-                </schedule-tick>
-                <schedule-event
-                    v-for="event in table.events"
-                    :key="
-                        $makeKey(
-                            event.event_id,
-                            'schedule_table_event',
-                            event.event_type,
-                        )
-                    "
-                    :value="event"
-                    :timeline-begin="table.timeline.begin"
-                ></schedule-event>
-            </schedule-table>
-            <schedule-list class="lg:hidden">
-                <schedule-list-group
-                    v-for="(group, i) in list.groups"
-                    :key="$makeKey(i, 'schedule_list_group')"
-                >
-                    <schedule-tick>{{ group.tick }}</schedule-tick>
-                    <schedule-event
-                        v-for="event in group.events"
-                        :key="$makeKey(event.event_id, 'schedule_list_event')"
-                        :value="event"
-                        in-list
-                    >
-                        <template #prepend>
-                            <schedule-room
-                                :value="event.room"
-                                inline
-                            ></schedule-room>
-                        </template>
-                    </schedule-event>
-                </schedule-list-group>
+            <schedule-grid :schedule="schedulesData[selectedDayIndex]">
+            </schedule-grid>
+            <schedule-list :schedule="schedulesData[selectedDayIndex]">
             </schedule-list>
         </div>
     </div>
@@ -72,13 +26,8 @@ import i18n from '@/i18n/conference/schedule.i18n'
 
 import CoreH1 from '@/components/core/titles/H1'
 import ScheduleDayTab from '@/components/schedule/ScheduleDayTab'
-import ScheduleEvent from '@/components/schedule/ScheduleEvent'
+import ScheduleGrid from '@/components/schedule/ScheduleGrid'
 import ScheduleList from '@/components/schedule/ScheduleList'
-import ScheduleListGroup from '@/components/schedule/ScheduleListGroup'
-import ScheduleRoom from '@/components/schedule/ScheduleRoom'
-import ScheduleRooms from '@/components/schedule/ScheduleRooms'
-import ScheduleTable from '@/components/schedule/ScheduleTable'
-import ScheduleTick from '@/components/schedule/ScheduleTick'
 
 export default {
     i18n,
@@ -86,13 +35,8 @@ export default {
     components: {
         CoreH1,
         ScheduleDayTab,
+        ScheduleGrid,
         ScheduleList,
-        ScheduleListGroup,
-        ScheduleRoom,
-        ScheduleRooms,
-        ScheduleTable,
-        ScheduleEvent,
-        ScheduleTick,
     },
     async asyncData({ store, payload }) {
         if (payload) return { schedulesData: payload }
@@ -110,34 +54,7 @@ export default {
     data() {
         return {
             selectedDayIndex: 0,
-            rooms: [],
-            tables: [],
-            lists: [],
-            defaultTable: {
-                events: [],
-                ticks: [],
-                timeline: {},
-            },
-            defaultList: {
-                groups: [],
-            },
         }
-    },
-    computed: {
-        table() {
-            return this.tables[this.selectedDayIndex] || this.defaultTable
-        },
-        list() {
-            return this.lists[this.selectedDayIndex] || this.defaultList
-        },
-    },
-    watch: {
-        selectedDayIndex() {
-            this.makeRooms()
-        },
-    },
-    created() {
-        this.processData()
     },
     activated() {
         setTimeout(() => {
@@ -152,82 +69,6 @@ export default {
             0
     },
     methods: {
-        processData() {
-            this.makeRooms()
-            this.makeTables()
-            this.makeLists()
-        },
-        makeRooms() {
-            this.rooms = ['4-r0', '5-r1', '6-r2', '1-r3']
-        },
-        makeTables() {
-            this.tables = this.schedulesData.map((schedule) => ({
-                events: this.getEvents(schedule),
-                ticks: this.getTicks(schedule),
-                timeline: schedule.timeline,
-            }))
-        },
-        makeLists() {
-            this.lists = this.schedulesData.map((schedule) => ({
-                groups: this.getGroups(schedule),
-            }))
-        },
-        getEvents(schedule) {
-            return schedule.rooms.reduce((events, room) => {
-                const slots = schedule.slots[room].map((slot) => ({
-                    ...slot,
-                    room,
-                }))
-                return [...events, ...slots]
-            }, [])
-        },
-        getTicks(schedule) {
-            const { timeline } = schedule
-            const start = this.$parseDate(timeline.begin)
-            const end = this.$parseDate(timeline.end)
-            const diff = end.diff(start, 'minute')
-            const unitInMinutes = 30
-            const unitPerGridColumn = 6
-            const tickCount = parseInt(`${diff / unitInMinutes}`, 10) + 1
-            return Array.from({ length: tickCount }).map((val, i) => {
-                const gridRowStart = unitPerGridColumn * i + 1
-                const gridRowEnd = unitPerGridColumn * i + unitPerGridColumn
-                const label = start
-                    .add(i * unitInMinutes, 'minute')
-                    .format('HH:mm')
-                const style = { gridRowStart, gridRowEnd }
-                return { label, style }
-            })
-        },
-        getGroups(schedule) {
-            const slotsByBeginTime = schedule.rooms.reduce((result, room) => {
-                schedule.slots[room].forEach((slot) => {
-                    const item = { ...slot, room }
-                    if (result[slot.begin_time]) {
-                        result[slot.begin_time].push(item)
-                    } else {
-                        result[slot.begin_time] = [item]
-                    }
-                })
-                return result
-            }, {})
-            return Object.keys(slotsByBeginTime)
-                .sort((timeA, timeB) => {
-                    if (timeA > timeB) {
-                        return 1
-                    } else if (timeA < timeB) {
-                        return -1
-                    } else {
-                        return 0
-                    }
-                })
-                .map((beginTime) => ({
-                    tick: this.$datetimeToString(beginTime, {
-                        outputFormat: 'HH:mm',
-                    }),
-                    events: slotsByBeginTime[beginTime],
-                }))
-        },
         selectDayTab(selectedIndex) {
             this.selectedDayIndex = selectedIndex
         },
